Move NoOpProgress out of test utils
