+++ betty/console/user.py
@@ -19,7 +19,7 @@ from betty.console.progress import ConsoleProgress
 from betty.console.rich import ConsoleTheme
 from betty.locale.localizable import Localizable
 from betty.progress import Progress
-from betty.test_utils.progress import NoOpProgress
+from betty.progress.no_op import NoOpProgress
 from betty.typing import Void, internal
 from betty.user import User, Verbosity
 from betty.user.logging import UserHandler
