+++ betty/tests/progress/test_no_op.py
@@ -1,0 +1,11 @@
+from betty.progress.no_op import NoOpProgress
+
+
+class TestNoOpProgress:
+    async def test_add(self) -> None:
+        sut = NoOpProgress()
+        await sut.add()
+
+    async def test_done(self) -> None:
+        sut = NoOpProgress()
+        await sut.done()
