+++ betty/user/no_op.py
@@ -11,8 +11,8 @@ from typing_extensions import override
 from betty.assertion import Assertion
 from betty.locale.localizable import Localizable
 from betty.progress import Progress
+from betty.progress.no_op import NoOpProgress
 from betty.test_utils.console import _T
-from betty.test_utils.progress import NoOpProgress
 from betty.typing import Void
 from betty.user import User, UserTimeoutError
